refactor(activerecord): move Migrator into migration.ts, create Store and IndexDefinition classes
- Migrator merged into migration.ts (Rails defines it in migration.rb)
- migrator.ts deleted, all imports updated
- Store class created in connection-adapters/abstract/query-cache.ts
  (extends QueryCacheStore)
- IndexDefinition promoted from interface to class in schema-definitions.ts
  with table, name, unique, columns fields matching Rails

api:compare: 59/597 (9.9%), 3 misplaced (was 6).
